agentic_tools/agent_totalsegmentator/requirements_minimal.txt
boto3==1.40.54
strands-agents==1.20.0
sentence-transformers==5.2.0
scikit-learn==1.8.0
spacy==3.8.11
TotalSegmentator==2.12.0
simpleitk==2.5.3
nibabel==5.3.3
numpy==2.3.1
torch==2.9.1